pyproject.toml
[project]
name = "iceberg-mcp-server"
version = "0.1.0"
description = "Cloudera Iceberg MCP Server via Impala"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "impyla>=0.21.0",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
]
